requirements.txt
Django==3.2.8
vaderSentiment==3.3.2
djangorestframework==3.13.1
whitenoise==5.3.0
gunicorn==20.1.0

